Add Bitwarden package, image, and `Log In` command
The `password-managers` package category has been added, because I
didn't see another category that felt appropriate.

Logging in authenticates with the Bitwarden server and saves a session
token to the user's default keychain. This enables the `Log In` script
command to also unlock a user's vault. For more information on the
Bitwarden CLI's authentication pattern, see:
https://bitwarden.com/help/article/cli/#session-management