Allow copying of a first matching password when multiple items are found
The `bw get item` command errors if the search query returns multiple
items. This change enables the intended behavior of copying the password
of the first matching item to the clipboard, by using `jq` to limit the
search results to the first item instead of the Bitwarden CLI.

Also explicitly `unset`s the `$password` value after is has been copied.