Add a Bitwarden package (#214)
* Add Bitwarden package, image, and `Log In` command

The `password-managers` package category has been added, because I
didn't see another category that felt appropriate.

Logging in authenticates with the Bitwarden server and saves a session
token to the user's default keychain. This enables the `Log In` script
command to also unlock a user's vault. For more information on the
Bitwarden CLI's authentication pattern, see:
https://bitwarden.com/help/article/cli/#session-management

* `Log In` script is `silent`, prettier documentation

* Add `Vault Status` script

* Add `Lock Vault` script

* Explicitly `unset` variables containing session tokens

This might improve the overall security of the package. If it doesn't,
it certainly doesn't hurt ¯\_(ツ)_/¯

* Add `Unlock` script

* Add `Log Out` script

* Better variable names and clearer spacing

* Add `Search Vault Items` script

* Add `Copy First Password` script

* Add `README.md`, supporting images

* Align script names with image names and documentation

* Provide a better error message if dependencies are missing

* Rename 'Copy First Password' command to 'Copy First Matching Password'

* Allow copying of a first matching password when multiple items are found

The `bw get item` command errors if the search query returns multiple
items. This change enables the intended behavior of copying the password
of the first matching item to the clipboard, by using `jq` to limit the
search results to the first item instead of the Bitwarden CLI.

Also explicitly `unset`s the `$password` value after is has been copied.

* Improve authentication and session token documentation

* Actually say what the `Log In` command does

* Omit hidden fields in search results by default

Vault items may contain hidden fields. The package should assume that
these fields contain sensitive data, and should treat them like
passwords.

* Update screenshot to match new command name

* Ignore useless `type` property of fields

* Use the long form of command args where possible