Create set-frontmost-as-default-browser
## Rationale
I see the "Switch Default Browser to ..." in the script list by @marcklingen, and would like to add my personal script that I've been using for 2 years. I created this back when I use Firefox, Brave, Safari and Safari Technology Preview(now not included in the script).

The implementation is similar, so I added the "... `defaultbrowser` needed" and "icon 🧭" for the sake of consistency.