translate-shell: Change setting for trans and gawk binaries
The current approach wasn't very clear how the PATH variable was used
later in the script.
Now the actual required binaries can be changed and are being used.

This also fixes an issue with the implementation which relied on GNU
tail and head which are by default not available on MacOS.
By hard coding the complete path ensure that BSD tail and head are being
used.