Update README.md's email address
Original link for feedback email was a link to `https://github.com/raycast/script-commands/feedback@raycast.com` which is not avlid.
Changed to `mailto:feedback@raycast.com` which is valid.

Another option is to present the link in a code block, that is `feedback@raycast.com`, which IMO is the better option since practically nobody really likes clicking `mailto:` links and has his email setup. I reckon most people that are inclined to send an email would copy the email address.