feat(browsing): command for doi string parsing
This is an additional command based on my previous
commit that scans the clipboard for doi strings
(rather than waiting for the user to paste the
string) and opens the corresponding resource.

Signed-off-by: Razvan Azamfirei <azamfireirazvan@gmail.com>

feat(browsing): command for doi string parsing

This is a package that parses a doi string (unique digital object identifier used to identify academic, professional, and government information) and opens the corresponding resource.