[script] Tailscale Switch: Make it switch account without logout
Update script tailscale-switch.sh to switch to another Tailscale account
without logging out of the current account.

With two logged-in Tailscale accounts, the updated script allows
switching between the two accounts without any re-authentication (as far
as tested for the duration of ~17 hours after login to my accounts).

The script uses `tailscale switch --list` to list the available accounts. It
takes the first not-connected account, and switches to this account with
`tailscale switch ID`.

```console
$ tailscale switch --list
ID    Tailnet                    Account
e163  tailnet-a                  account-a@tailnet-a.example
2b13  tailnet-b                  account-b@tailnet-b.example
```

```console
$ tailscale switch 2b13
Switching to account "2b13"
Success.
```